guard against removing extra related records (#22)
when removing a relatedRecord, if it was not found it would remove the
relationship for the last record (at index -1). This had the side effect
of removing two records when one was found because of the recursive
nature between `remove` and `setRelatedRecord`.

also, record ids are integers but relationship ids are strings, so were
failing to remove relationship data for objects with persistent ids. There
is a more comprehensive fix for this coming in
https://github.com/artemis-ag/mobx-async-store/pull/27 but this slaps
a quick string coercion into our checks for the time being.