fix bug cloning relationships
we want the relationships themselves to be cloned, but of course we want
the targets of the relationships (the related objects) to not be cloned
and to be referenced instead. I think I thought I was doing this when I
wrote Model#clone originally, it looks like I tried to write a test for
it too, but the behavior I was after was already achieved via object
proxies, so no special treatment was needed to avoid cloning the related
objects themselves. I added one more test here to further clear up the
distinction.