flatten all error keys and add full server data
mobx-async-store is designed to work to the spec and with jsonapi-rails
so the errors should be preserved in their entirety.

Since error keys are already flattened in the rails errors it is
appropriate to maintain that strucutre when parsing errors. This way we
don't have to worry about how deeply structured the errors are or how
they will be used. Individual Models should add helpers to parse errors
in a more straightforward way that is particular to the data.