5.0.0 (#401)
Consistently handle errors between read/write/deletes. When any of these operations fail they
will return a rejected promise with an error. That error's message will be a JSON stringified array
of JSONAPI error objects. In the case of a server error with no parsable errors, mobx-async-store
will build a JSONAPI error object based on the status of the response and a new `errorMessages`
configuration.