fix: removed duplicate records from getAll (#489)
in order to preserve relationships that may be pointing to a temp id
from a new record after that record is saved, the store's data contains
a duplicate reference to the record - one from its temp id and one from
its persisted id. When retrieving these values via `getAll`, we should
only return that record once.