return jsonapi version in MockServer responses
this version is configurable when creating the mock server via the
`jsonapiVersion` option. It will only be included in responses
leveraging the backend store - if you are using responseOverrides, then
you should provide the version in your mocked response as needed. This
enables mixed version responses in a single MockServer (e.g. if you are
migratiing some endpoints to another version).

if using the provided `serverResponse` utility to compose your response
overrides, you can pass the version in via a new options parameter

```js
mockServer.responseOverrides = [
  {
    path: '/users/1',
    method: 'GET',
    response: (mockServer) => serverResponse(mockServer.store.getOne('users', 1), { version: '1.0' }),
  }
]
```